+++ chainlibpy/amino/__init__.py
@@ -1,7 +1,20 @@
-from .basic import (DEFAULT_BECH32_HRP_BASE, Coin, CommissionRates, Content,
-                    Description, Input, Output, StdFee, SyncMode,
-                    TimeoutHeight, VoteOptionAbstain, VoteOptionNo,
-                    VoteOptionNoWithVeto, VoteOptionUnspecified, VoteOptionYes)
+from .basic import (
+    DEFAULT_BECH32_HRP_BASE,
+    Coin,
+    CommissionRates,
+    Content,
+    Description,
+    Input,
+    Output,
+    StdFee,
+    SyncMode,
+    TimeoutHeight,
+    VoteOptionAbstain,
+    VoteOptionNo,
+    VoteOptionNoWithVeto,
+    VoteOptionUnspecified,
+    VoteOptionYes,
+)
 from .signdoc import StdSignDoc
 
 __all__ = [